+++ index.js
@@ -1,0 +1,2 @@
+export { usePaging } from "./src/ss-paging-esm";
+export { usePagingStore } from './src/ss-paging-store'
