+++ module.d.ts
