+++ src/ss-paging-store.ts
@@ -1,4 +1,4 @@
 import { defineStore } from 'pinia'
 import { usePaging } from './ss-paging-esm'
 
-export const usePagingStore = defineStore('paging', () => usePaging())
+export const usePagingStore = defineStore('sspaging', () => usePaging())
